refactor : [Step2] 인터페이스 분리 및 클래스 명 구체화
