+++ src/main/java/nextstep/subway/path/domain/Path.java
@@ -1,42 +1,13 @@
 package nextstep.subway.path.domain;
 
-import java.util.Collections;
 import java.util.List;
 
-import org.jgrapht.GraphPath;
-
-import nextstep.subway.exception.path.PathException;
 import nextstep.subway.station.domain.Station;
 
-public class Path {
+public interface Path {
 
-    private List<Station> stations;
-    private List<SectionEdge> sectionEdges;
+    public int getMinDistance();
 
-    public Path(List<Station> stations, List<SectionEdge> sectionEdges) {
-        this.stations = stations;
-        this.sectionEdges = sectionEdges;
-    }
-
-    public int getMinDistance() {
-        return sectionEdges.stream()
-            .mapToInt(SectionEdge::getDistance)
-            .sum();
-    }
-
-    public List<Station> getStations() {
-        return Collections.unmodifiableList(stations);
-    }
-
-    public static Path of(GraphPath<Station, SectionEdge> path) {
-        validatoin(path);
-        return new Path(path.getVertexList(), path.getEdgeList());
-    }
-
-    private static void validatoin(GraphPath<Station, SectionEdge> path) {
-        if (path == null) {
-            throw new PathException(PathException.NOT_CONNECTED);
-        }
-    }
+    public List<Station> getStations();
 
 }
